Release/v1.8.6 x

+++ libs/model/src/contest/CancelContestManagerMetadata.command.ts
@@ -23,7 +23,11 @@ export function CancelContestManagerMetadata(): Command<
       });
       mustExist('Contest Manager', contestManager);
 
-      if (contestManager.neynar_webhook_id) {
+      // delete webhook *only* if it was created on the current environment (e.g. local, beta, prod)
+      if (
+        contestManager.environment === config.APP_ENV &&
+        contestManager.neynar_webhook_id
+      ) {
         const client = new NeynarAPIClient(config.CONTESTS.NEYNAR_API_KEY!);
         try {
           await client.deleteWebhook(contestManager.neynar_webhook_id);
